requirements.txt
opencv_python==4.9.0.80
Pillow==10.3.0
protobuf==3.20.3
pytesseract==0.3.10
streamlit==1.22.0
